requirements.txt
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.2
dnspython==2.5.0
email-validator==2.0.0.post2
exceptiongroup==1.2.0
fastapi==0.109.0
fastapi-cache2==0.2.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
flower==2.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.5
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.4
orjson==3.9.12
packaging==23.2
passlib==1.7.4
pendulum==3.0.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
pydantic==2.6.0
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
PyJWT==2.8.0
pytest==7.4.4
pytest-asyncio==0.23.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.6
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.13.0
tomli==2.0.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
uvicorn==0.27.0.post1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
flower
gunicorn